refactor: 핵심 비지니스 로직을 domain 패키지로 이동
