+++ src/main/java/study/racing/domain/CarRacing.java
@@ -1,4 +1,4 @@
-package study.racing;
+package study.racing.domain;
 
 import study.racing.view.InputView;
 import study.racing.view.ResultView;
@@ -25,16 +25,5 @@ public class CarRacing {
             racingGameRounds.recording(cars);
         }
         return racingGameRounds.getRecordingResult();
-    }
-
-    public static void main(String args[]){
-        String[] carNames = InputView.getCarsName();
-        int tryCnt = InputView.getTryCount();
-
-        CarRacing carRacing = new CarRacing(new RandomMoveStrategy());
-        RacingRecords racingResults = carRacing.start(carNames, tryCnt);
-
-        ResultView.showResultHead();
-        ResultView.showResult(racingResults);
     }
 }
