+++ src/main/java/study/racing/domain/RacingErrorMessage.java
@@ -1,4 +1,4 @@
-package study.racing;
+package study.racing.domain;
 
 import study.racing.view.InputView;
