+++ src/main/java/study/racing/domain/RacingMoveStrategy.java
@@ -1,4 +1,4 @@
-package study.racing;
+package study.racing.domain;
 
 public interface RacingMoveStrategy {
     int getMovement();
