+++ src/main/java/study/racing/domain/RacingRecord.java
@@ -1,4 +1,4 @@
-package study.racing;
+package study.racing.domain;
 
 import java.util.LinkedList;
