+++ src/test/java/study/racing/RacingGameRoundsTest.java
@@ -4,6 +4,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import study.racing.domain.*;
+
 import java.util.LinkedList;
 import static org.junit.jupiter.api.Assertions.assertEquals;
