+++ src/test/java/study/racing/RacingInputValidationTest.java
@@ -3,6 +3,9 @@ package study.racing;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import study.racing.domain.CarRacing;
+import study.racing.domain.RacingErrorMessage;
+import study.racing.domain.RandomMoveStrategy;
 import study.racing.view.InputView;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
